+++ homework-4/main.py
@@ -13,5 +13,5 @@ if __name__ == '__main__':
     assert item1 + phone1 == 25
     assert phone1 + phone1 == 10
 
-    phone1.number_of_sim = 0
+    #phone1.number_of_sim = 0
     # ValueError: Количество физических SIM-карт должно быть целым числом больше нуля.
